requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via httpx
attrs==24.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via
    #   pdfminer-six
    #   requests
click==8.1.8
    # via streamlit
colorama==0.4.6
    # via click
cryptography==44.0.0
    # via pdfminer-six
dataclasses-json==0.6.7
    # via langchain-community
faiss-cpu==1.9.0.post1
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via langsmith
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
langchain==0.3.13
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.13
    # via -r requirements.in
langchain-core==0.3.28
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.3.4
    # via langchain
langsmith==0.2.6
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.23.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
narwhals==1.19.1
    # via altair
numpy==2.2.1
    # via
    #   faiss-cpu
    #   langchain
    #   langchain-community
    #   pandas
    #   pydeck
    #   streamlit
orjson==3.10.12
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   faiss-cpu
    #   langchain-core
    #   marshmallow
    #   streamlit
pandas==2.2.3
    # via streamlit
pdfminer-six==20231228
    # via pdfplumber
pdfplumber==0.11.4
    # via -r requirements.in
pillow==11.0.0
    # via
    #   pdfplumber
    #   streamlit
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.2
    # via streamlit
pyarrow==18.1.0
    # via streamlit
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.0
    # via langchain-community
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pymupdf==1.25.1
    # via -r requirements.in
pypdf==5.1.0
    # via -r requirements.in
pypdfium2==4.30.1
    # via pdfplumber
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via
    #   langchain
    #   langchain-community
streamlit==1.41.1
    # via
    #   -r requirements.in
    #   streamlit-chat
streamlit-chat==0.1.1
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
watchdog==6.0.0
    # via streamlit
yarl==1.18.3
    # via aiohttp
